【Issue：#1】整數相除
1. 測試類別：DivisionTests
    * 建立測試案例：Divide_WhenDividendIsThreeDivisorIsTwo_ReturnsOnePointFive
2. 實作類別：Division
    * 調整方法實作：Divide

+++ TDD.Arithnetic/Division.cs
@@ -4,7 +4,7 @@
     {
         public static decimal Divide(int dividend, int divisor)
         {
-            decimal quotient = dividend / divisor;
+            decimal quotient =(decimal) dividend / (decimal)divisor;
 
             return quotient;
         }
